+++ core/src/main/java/org/msh/pharmadex/dao/UserDAO.java
@@ -1,7 +1,7 @@
 package org.msh.pharmadex.dao;
 
 import org.hibernate.Hibernate;
-import org.msh.pharmadex.domain.*;
+import org.msh.pharmadex.domain.User;
 import org.msh.pharmadex.domain.enums.UserType;
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
@@ -10,11 +10,9 @@ import javax.persistence.EntityManager;
 import javax.persistence.NoResultException;
 import javax.persistence.PersistenceContext;
 import javax.persistence.PersistenceContextType;
-import javax.persistence.criteria.*;
 import java.io.Serializable;
 import java.util.Date;
 import java.util.List;
-import java.util.Map;
 
 /**
  * Created by IntelliJ IDEA.
@@ -30,7 +28,7 @@ public class UserDAO implements Serializable {
     EntityManager entityManager;
 
     @Transactional
-    public User findUser(int id) {
+    public User findUser(Long id) {
 
         User user = (User) entityManager.createQuery("select u from User u where u.userId = :userid")
                 .setParameter("userid", id)
